Separated UI and backend, removed User Interface directory
