--- src/backend/Networking/test/index.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Beaglebone Controller</title>
-
-    <style>
-        body {
-            text-align: center;
-            margin: 0px;
-            padding: 0px;
-        }
-    </style>
-    <script src ="Beaglebone_Controller.js" defer></script>
-</head>
-<body>
-    <canvas id="gaming zone"></canvas>
-    <p>[Put Controller Image Here]</p>
-</body>
-</html>
