--- src/frontend/UI/index.html
@@ -1,47 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-	<script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/4.0.1/socket.io.js" integrity="sha512-q/dWJ3kcmjBLU4Qc47E4A9kTB4m3wuTY7vkFJDTZKjTs8jhyGQnaUrxa0Ytd0ssMZhbNua9hE+E7Qv1j+DyZwA==" crossorigin="anonymous"></script>
-	
-    <title>Beaglebone User Interface</title>
-
-    <style>
-      body {
-        text-align: center;
-        margin: 0px;
-        padding: 0px;
-      }
-
-      /*rectangular guage to track how fast car is accelerating/decelerating*/
-      .acceleration_guage {
-        color: black;
-        border: black;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        position: relative;
-        width: 0;
-        height: 0;
-      }
-
-      /*circular guage to representation direction in which car is being steered (compass-like)*/
-      .turn_guage {
-        display: flex;
-        justify-content: left;
-        align-items: center;
-        height: 50vh;
-      }
-    </style>
-
-	<script type="text/javascript"
-		src="{{ url_for('static', filename='Beaglebone_Controller.js') }}"></script>
-  </head>
-  <body>
-    <div class="background">
-      <div class="acceleration_guage"></div>
-    </div>
-  </body>
-</html>
